Automatic Image Thumbnail Generator/requirements.txt
flask
boto3
python-dotenv
Pillow